Commit Week 15/05-19/05 	-Refactor liquec to add big decimal 	-Added web converters 	-First approach berock front end 	-Refactor literals 	-Finished berock result view 	-Fix berock table 	-Added formatter number 	-Finished create berock 	-Added liquibase 	-Added module-berock to Dockerfile 	-Fix H2 issue

+++ module-liquec/src/main/java/edu/uoc/geopocket/liquec/calculation/tasks/EarthquakeMagnitudeCorrectionLiquecTask.java
@@ -8,6 +8,8 @@ import edu.uoc.geopocket.liquec.entities.Spt;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Slf4j
 @Component
 public class EarthquakeMagnitudeCorrectionLiquecTask extends AbstractLiquecTaskExecutable {
@@ -18,7 +20,8 @@ public class EarthquakeMagnitudeCorrectionLiquecTask extends AbstractLiquecTaskE
 
     public void execute(final Liquec liquec, final Spt targetSpt) {
 
-        final Double earthquakeMagnitudeCorrection = Polynomial.EARTHQUAKE_MAGNITUDE_CORRECTION.getValue(liquec.getEarthquakeMagnitude());
+        final BigDecimal earthquakeMagnitudeCorrection =
+                Polynomial.EARTHQUAKE_MAGNITUDE_CORRECTION.getValue(BigDecimal.valueOf(liquec.getEarthquakeMagnitude()));
 
         log.debug("Earthquake magnitude correction (CM): " + earthquakeMagnitudeCorrection);
